rag_service/requirements.txt
fastapi
uvicorn
langchain
langchain-community
faiss-cpu
openai
python-dotenv
sqlalchemy
